requirements.txt
fastapi
uvicorn
pydantic
python-multipart
speechrecognition
pyttsx3
gTTS
python-dotenv
pydub
soundfile
pocket-tts